# use fixed pool to improve perf of aead ChaCha20Poly1305