~35-45% speedup on benchmarks when reusing the same key for multiple operations.

Other AEADs will follow, although they're...harder.

```
| tests/bench/test_aead.py::test_chacha20poly1305_decrypt | 7125.08 ns | 4375.12 ns | 39% faster |
| tests/bench/test_aead.py::test_chacha20poly1305_encrypt | 8500.18 ns | 4709.0 ns | 45% faster |
```