@reaperhulk 2022-09-12
@bdraco last year you talked to us about speeding up ChaCha20Poly1305 by ~20% when reusing the same key repeatedly. IIRC the example use case was home-assistant HomeKit integration since HAP uses ChaCha20Poly1305. We've just landed this PR, which (using the exact same API and in a thread safe manner) allows ctx reuse with 35%+ performance improvement and it will be released in version 39. I'll send a PR to home-assistant after release, but I thought you'd likely be interested. 😄