@bdraco 2022-09-12
Hi Paul,

Thanks for the ping.  This is fantastic news!

Our solution was to create https://github.com/bdraco/chacha20poly1305-reuseable 

Its currently used in https://github.com/ikalchev/HAP-python https://github.com/postlund/pyatv and https://github.com/Jc2k/aiohomekit

There was one other change since the original PR to squeeze a bit more performance out
https://github.com/bdraco/chacha20poly1305-reuseable/commit/be14cc2dc74d81b49bf761c32ce5536e31be74ce#diff-61eb0fe778bd395a16e1d556d37bc4cc2c1649b88e9d797eeee8f839125dfdf6R114

I'm hoping we can drop that lib and switch everything back over to use your changes.

Cheers
-Nick