@reaperhulk 2022-09-12
@bdraco Oh, fascinating. I dropped your code into my test harness and (unsurprisingly) you're faster since you don't have the locking synchronization overhead:

```
| tests/bench/test_aead.py::test_chacha20poly1305_decrypt | 4583.04 ns | 3458.0 ns | 25% faster |
| tests/bench/test_aead.py::test_chacha20poly1305_encrypt | 4749.98 ns | 4374.89 ns | 8% faster |
```

But the original timings were 7100ns and 8500ns, respectively, so we're hopefully performant enough that you can drop this 😄